+++ Code/Client/users/src/components/molecules/signup/signupModel.module.css
@@ -1,0 +1,88 @@
+.modal_content {
+    z-index: 1100;
+    top: 55%; /* Vertically center */
+    left: 50%; /* Horizontally center */
+    transform: translate(-50%, -50%);
+    background-color: white;
+    position: fixed;
+    padding: 0px;
+    width: 900px;
+    height: 500px;
+    border-radius: 30px;
+    transition: 0.5s ease-in;
+    display: flex;
+  }
+  .login_form_container {
+    padding: 12px 40px;
+    width: 580px;
+    border-top-right-radius: 30px;
+    border-bottom-right-radius: 30px;
+    background-color: #b491d2;
+  }
+  .login_form_container > form {
+    color: white;
+    display: flex;
+    flex-direction: column;
+    align-items: stretch;
+  }
+  .image {
+    flex-grow: 1;
+  }
+  .image img{
+    width: 20.9rem; 
+    height: 500px; 
+    border-top-left-radius: 30px;
+    border-bottom-left-radius: 30px;
+  }
+  .heading{
+    font-size: 1.6rem;
+  }
+  .row {
+    display: flex;
+    gap: 2.5rem;
+}
+  .login_form > div {
+    margin: 20px 0px;
+    font-weight: 600;
+    letter-spacing: -2px;
+    text-align: center;
+    border-radius: 5px;
+    text-transform: uppercase;
+  }
+
+  .inputs, .teaxtarea{
+    padding: 7px 14px;
+    font-size: 1.1rem;
+    border-radius: 5px;
+    border: none;
+  }
+
+  .teaxtarea{
+   height: 5rem;
+   width: 31rem;
+  }
+
+  .button {
+    margin: 20px 0px;
+    background-color:  #8d48c9;
+    color: white;
+    padding: 10px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+  }
+  
+  
+  @media screen and (max-width:1068px) {
+    .modal-content{
+        height: 100% ;
+        width: 100%;
+        justify-content: center;
+        align-items: center;
+        background-image: none;
+        border-radius: 0px;
+        /* padding: 20p; */
+        margin: 0;
+  
+    }
+  }
