+++ Code/Client/users/src/pages/signUp/signup.module.css
@@ -1,0 +1,5 @@
+  .signup {
+    position: fixed;
+    inset: 0;
+    z-index: 100;
+  }
